Extensive Tests, Bug Fixes, Performance Work, Initial Wiki (#56)
* Be more allocation-friendly, expose more helpers, be more SINGLE_THREADED friendly

* Test fixes, bug fixes

* Even more tests, performance enhancements

* Many fixes

* More tests

* Add string comparisons

* Add functionality to Child/Sibling/Parent components

* Minor spatial tree bug fixes + extensive tests

* CyclicBuffer bug fixes

* Reduce allocations in LineHelpers, fix minor bugs

* Minor bug fixes in Helpers, more tests

* Remove regions

* Many test and prod fixes

* Fix all tests

* Relational component performance

* Optimizations

* RelationalComponent cleanups

* Better AsyncOperations

* Progress on tests

* Enhanced Singletons

* Fix remaining tests

* Update spatial tree performance

* Performance optimizations in RTree

* RTree construction optimization

* Simplification of for loops

* Optimize QuadTree

* Further KDTree Improvements

* PR feedback

* Benchmarks auto-populate readme

* Update SpatialTree -> SpatialTree2D

* Improved KDTree construction

* In-Progress

* In-Progress

* Simplify reflection

* RTree updates

* Add destruction buffers

* CyclicBuffer perf improvements, RTree perf improvements

* Even more RTree2D optimizations

* Initial 3D spatial trees

* Minor updates

* Problems with tests

* Member renames

* Fix labeling

* Progress

* Update README with new perf numbers

* Progress

* Update Random performance

* Many bug fixes

* RTree3D fixes

* Update DataStructures

* Add additional IList tests

* Add ImmutableBitSet tests

* Fix many casing issues

* Remove debug logging from CyclicBuffer

* Full test coverage

* Simplify ProtoEquals

* Fix Sphere/Circle

* Update Sphere/SphereTests

* Progress

* Fix Polygon tests

* Remove temp files

* Remove package lock files

* Improve performance of helpers using caching/pools

* Runtime caching

* Updates

* More faster relational components

* More performance in Relational Components

* Fix PolygonCollider2D editor not tracking changes

* More updates

* Fix memory allocations in Enum

* Performance for RelationalComponents

* Progress

* Update randoms

* Fix compilation

* Updates

* Progress

* Fixes

* Updates

* TextureScale updates

* Progress

* Progress

* README updates

* Progress

* Line3D fixes

* Updates

* Updates

* Fixes

* Progress

* SpatialTree updates

* Progress

* Editor Tooling Updates

* Progress

* Spatial Tree Updates

* SpatialTree Updates

* More Spatial Tree Updates

* Updates

* README update

* Fixes

* README updates

* Updates

* Fixes

* Updates

* LZMA bugfix

* Remove unused imports

* Fixes

* Fix RelationalTest issues

* Test updates

* Common Test Base

* Common Test base

* Meta files

* Common test progress

* Updates

* Common Test Usage

* Warning fixes

* README updates

* Doc updates

* Fix RuntimeSingletonTests

* Update THIRD PARTY NOTICES

* THIRD PARTY NOTICES updates

* Progress

* Progress

* Update SpriteSettingsApplierAPI

* Editor fixes

* Fixes

* Hull fixes

* README extensions

* Updates

* Updates

* Progress

* Updates

* Progress

* PRoto fixes

* Fix WyRandom serialization

* Fix logging in AnimationCopier

* Serialization Fixes

* Progress

* Update Serialization docs

* Serializer updates

* Proto fixes

* Create REFLECTION_HELPERS meta file

* Fixes

* JSON updates

* JSON updates

* README updates

* Improvements

* Fixes

* Test fixes

* Updates

* Update

* Spatial Tree Updates

* More docs

* Progress

* Progress

* Progress

* SVG updates

* More SVG updates

* SVG updates

* Updates

* Fixes

* SVG fixes

* SVG updates

* Updates

* Fix

* SVG updates

* Better geometry edge cases

* Better geometry intersection

* Better KD Tree

* Better KDTree

* Better KDTree

* Better KD Tree

* Better tree diagrams

* Better SVGs

* Fixes

* better OctTree

* Better KDTree

* Better RTree3D

* Clearer boundary semantics

* random_generators update

* Fix reflection scan

* Update spatial tree semantics

* Fix random generators

* Fixes

* Better relational wiring

* Relational fix

* Update relational wiring

* Fixes

* SVG updates

* SVG updates

* More SVGs

* better SVG

* Updates

* Progress

* Progress

* Better dequeue SVG

* SVG

* SVG finalization

* Updates

* README updates

* ReflectionHelper updates

* Documentation updates

* More ReflectionHelper tests

* doc updates

* Better EditorUi

* doc updates

* Update documentation links

* Fixes

* Test updates

* ReflectionHelper updates

* Progress

* FastVector JSON serialization issues

* Many JSON fixes

* Many proto/JSON serialization errors

* Finalized updates

* Update version

+++ Runtime/Utils/SevenZip/Compress/LZMA/LzmaBase.cs
@@ -77,7 +77,7 @@ namespace SevenZip.Compression.LZMA
                 return len;
             }
 
-            return (uint)(kNumLenToPosStates - 1);
+            return kNumLenToPosStates - 1;
         }
 
         public const int kNumAlignBits = 4;
